+++ packages/react-dictate-button/__tests__/multipleInterims.spec.tsx
@@ -96,7 +96,9 @@ describe('with multiple non-finalized interims', () => {
     });
 
     expect(onDictate).toHaveBeenCalledTimes(0);
+    expect(onEnd).toHaveBeenCalledTimes(0);
     expect(onProgress).toHaveBeenCalledTimes(1);
+    expect(onStart).toHaveBeenCalledTimes(0);
     expect(onProgress.mock.calls[0][0]).toHaveProperty('type', 'progress');
     expect(onProgress.mock.calls[0][0]).toHaveProperty('results', [
       { confidence: 0.009999999776482582, transcript: 'test' }
@@ -119,7 +121,9 @@ describe('with multiple non-finalized interims', () => {
     });
 
     expect(onDictate).toHaveBeenCalledTimes(0);
+    expect(onEnd).toHaveBeenCalledTimes(0);
     expect(onProgress).toHaveBeenCalledTimes(1);
+    expect(onStart).toHaveBeenCalledTimes(0);
     expect(onProgress.mock.calls[0][0]).toHaveProperty('type', 'progress');
     expect(onProgress.mock.calls[0][0]).toHaveProperty('results', [
       { confidence: 0.009999999776482582, transcript: 'testing' }
@@ -142,7 +146,9 @@ describe('with multiple non-finalized interims', () => {
     });
 
     expect(onDictate).toHaveBeenCalledTimes(0);
+    expect(onEnd).toHaveBeenCalledTimes(0);
     expect(onProgress).toHaveBeenCalledTimes(1);
+    expect(onStart).toHaveBeenCalledTimes(0);
     expect(onProgress.mock.calls[0][0]).toHaveProperty('type', 'progress');
     expect(onProgress.mock.calls[0][0]).toHaveProperty('results', [
       { confidence: 0.8999999761581421, transcript: 'testing' }
@@ -166,7 +172,9 @@ describe('with multiple non-finalized interims', () => {
     });
 
     expect(onDictate).toHaveBeenCalledTimes(0);
+    expect(onEnd).toHaveBeenCalledTimes(0);
     expect(onProgress).toHaveBeenCalledTimes(1);
+    expect(onStart).toHaveBeenCalledTimes(0);
     expect(onProgress.mock.calls[0][0]).toHaveProperty('type', 'progress');
     expect(onProgress.mock.calls[0][0]).toHaveProperty('results', [
       { confidence: 0.8999999761581421, transcript: 'testing' },
@@ -191,7 +199,9 @@ describe('with multiple non-finalized interims', () => {
     });
 
     expect(onDictate).toHaveBeenCalledTimes(0);
+    expect(onEnd).toHaveBeenCalledTimes(0);
     expect(onProgress).toHaveBeenCalledTimes(1);
+    expect(onStart).toHaveBeenCalledTimes(0);
     expect(onProgress.mock.calls[0][0]).toHaveProperty('type', 'progress');
     expect(onProgress.mock.calls[0][0]).toHaveProperty('results', [
       { confidence: 0.8999999761581421, transcript: 'testing' },
@@ -216,7 +226,9 @@ describe('with multiple non-finalized interims', () => {
     });
 
     expect(onDictate).toHaveBeenCalledTimes(0);
+    expect(onEnd).toHaveBeenCalledTimes(0);
     expect(onProgress).toHaveBeenCalledTimes(1);
+    expect(onStart).toHaveBeenCalledTimes(0);
     expect(onProgress.mock.calls[0][0]).toHaveProperty('type', 'progress');
     expect(onProgress.mock.calls[0][0]).toHaveProperty('results', [
       { confidence: 0.8999999761581421, transcript: 'testing' },
@@ -241,7 +253,9 @@ describe('with multiple non-finalized interims', () => {
     });
 
     expect(onDictate).toHaveBeenCalledTimes(0);
+    expect(onEnd).toHaveBeenCalledTimes(0);
     expect(onProgress).toHaveBeenCalledTimes(1);
+    expect(onStart).toHaveBeenCalledTimes(0);
     expect(onProgress.mock.calls[0][0]).toHaveProperty('type', 'progress');
     expect(onProgress.mock.calls[0][0]).toHaveProperty('results', [
       { confidence: 0.8999999761581421, transcript: 'testing' },
@@ -267,7 +281,9 @@ describe('with multiple non-finalized interims', () => {
     });
 
     expect(onDictate).toHaveBeenCalledTimes(1);
+    expect(onEnd).toHaveBeenCalledTimes(0);
     expect(onProgress).toHaveBeenCalledTimes(0);
+    expect(onStart).toHaveBeenCalledTimes(0);
     expect(onDictate.mock.calls[0][0]).toHaveProperty('type', 'dictate');
     expect(onDictate.mock.calls[0][0]).toHaveProperty('result', {
       confidence: 0.5359774827957153,